Moved confirm dialog to namespace also obsoleted container.
