+++ Core/Menu/IGMData/Obsolete/IGMData_Container.cs
@@ -1,8 +1,11 @@
-﻿namespace OpenVIII
+﻿using System;
+
+namespace OpenVIII
 {
     /// <summary>
     /// Contains only one IGMDataItem.
     /// </summary>
+    [Obsolete("This method will soon be deprecated. You can use IGMDataItem/IGMData directly. No need for this container Wrapper.")]
     public class IGMData_Container : IGMData.Base
     {
         #region Constructors
